Added QMA9681 register values, proper init
Modified MMA8452Q to return the same values as QMA6981
Properly datatyped acceleration values

+++ firmware/user/MMA8452Q.h
@@ -10,9 +10,9 @@
 
 typedef struct
 {
-    int x;
-    int y;
-    int z;
+    int16_t x;
+    int16_t y;
+    int16_t z;
 } accel_t;
 
 bool MMA8452Q_setup(void);
